creazione metodo createTask non funzionante

+++ src/main/java/it/intesys/codylab/repository/SlotRepository.java
@@ -3,7 +3,7 @@ package it.intesys.codylab.repository;
 import it.intesys.codylab.model.Slot;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SlotRepository extends CrudRepository<Slot, Integer> {
+public interface SlotRepository extends CrudRepository<Slot, Long> {
 
 
 }
