+++ src/main/java/it/intesys/codylab/repository/TaskRepository.java
@@ -3,6 +3,5 @@ package it.intesys.codylab.repository;
 import it.intesys.codylab.model.Task;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TaskRepository extends CrudRepository<Task, Integer> {
-
+public interface TaskRepository extends CrudRepository<Task, Long> {
 }
